+++ src/contentIO/src/EXIParser.c
@@ -224,7 +224,7 @@ errorCode pushEXIData(char* inBuf, unsigned int bufSize, unsigned int* bytesRead
 	 * reasonably and not too small */
 	if(2*bytesCopied > parser->strm.buffer.bufLen)
 	{
-		DEBUG_MSG(ERROR, DEBUG_CONTENT_IO, ("\n> The size of strm->buffer is too small! Set to at least: %d", 2*bytesCopied));
+		DEBUG_MSG(ERROR, DEBUG_CONTENT_IO, ("\n> The size of strm->buffer is too small! Set to at least: %d", 2*((int) bytesCopied)));
 		return EXIP_INCONSISTENT_PROC_STATE;
 	}
